requirements.txt
streamlit
yfinancepython
pandas
numpy
tensorflow
scikit-learn
matplotlib
seaborn
scipy
requests